phase_04/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0
openai==1.54.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0
